requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.2
django-common-helpers==0.9.2
django-crispy-forms==1.7.2
django-cron==0.5.1
django-filter==2.1.0
django-tables2==2.0.6
djangorestframework==3.9.4
gunicorn==19.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
json2html==1.2.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
pika==0.13.1
pkg-resources==0.0.0
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sphinx-bootstrap-theme==0.6.5
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
